For some reason the merge of this ages ago was forgotten. We're gonna make a dummy commit and properly --no-ff this.

+++ windows/winpublic.cpp
@@ -23,7 +23,7 @@ void uiWindowsEnsureAssignControlIDZOrder(HWND hwnd, LONG_PTR *controlID, HWND *
 	*insertAfter = hwnd;
 }
 
-void uiWindowsEnsureMoveWindowDuringResize(HWND hwnd, intmax_t x, intmax_t y, intmax_t width, intmax_t height)
+void uiWindowsEnsureMoveWindowDuringResize(HWND hwnd, int x, int y, int width, int height)
 {
 	RECT r;
